Hack to make code generation work with new trace metamodels
With new trace metamodels, generic type parameters are used.

For some reason, the former code to generated code did not work anymore
with these metamodels, and errors appeared in the generated code.

However, using the UI to generate the code from the trace metamodel
genmodel worked, mysteriously, for some reason.

Hence this commit fixes the problem by creating a fake GenModelEditor UI
to load the model, and to generate the code.

This is horrible, but it works. Sorry.