[DefaultModelLoader] Delay the ResourceSet configuration
Previously, when Melange was solving a cross-ref of the
model, a Melange query was always added to the URI of
the cross-ref although Melange needed to access to
the resource without any query.

This PR delays the configuration of the ResourceSet with
the XMLURIHandler and the MelangeURIConverter, until
after the model was entirely loaded, which includes
the MelangeResource creation, and avoids the aforementioned
problem.